+++ challenge-estoa/src/utils/get-image-path.tsx
